More cleanups to make golangci-lint happier
Moving some functionality from requester to case.

Lots of cleanup remains, so turning off some of pre-commit.
