+++ requester.go
@@ -3,12 +3,8 @@ package gobbi
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"io"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -54,57 +50,6 @@ func NewClient(ctx context.Context) *BaseClient {
 	return &b
 }
 
-func (b *BaseClient) updateQueryString(c *Case, u string) (string, error) {
-	additionalValues := c.QueryParameters
-	if len(additionalValues) == 0 {
-		// No changes required, return early
-		return u, nil
-	}
-	parsedURL, err := url.Parse(u)
-	if err != nil {
-		return u, fmt.Errorf("unable to parse url: %s: %w", u, err)
-	}
-	currentValues := parsedURL.Query()
-	for k, v := range additionalValues {
-		switch x := v.(type) {
-		case []interface{}:
-			s := make([]string, len(x))
-			for i, item := range x {
-				s[i] = scalarToString(item)
-			}
-			currentValues[k] = s
-		default:
-			currentValues[k] = []string{scalarToString(x)}
-		}
-	}
-	for k, vList := range currentValues {
-		for i, v := range vList {
-			newV, err := StringReplace(c, v)
-			if err != nil {
-				c.Errorf("unable to string replace query parameter %s: %v", k, err)
-				continue
-			}
-			currentValues[k][i] = newV
-		}
-	}
-
-	parsedURL.RawQuery = currentValues.Encode()
-	return parsedURL.String(), nil
-}
-
-func scalarToString(v any) string {
-	var sValue string
-	switch x := v.(type) {
-	case string:
-		sValue = x
-	case int:
-		sValue = strconv.Itoa(x)
-	case float64:
-		sValue = strconv.FormatFloat(x, 'G', -1, 64)
-	}
-	return sValue
-}
-
 // Do executes the current Case, first checking to see if it has any priors that
 // have not been executed.
 func (b *BaseClient) Do(c *Case) {
@@ -112,31 +57,12 @@ func (b *BaseClient) Do(c *Case) {
 	if c.Done() {
 		c.GetTest().Logf("returning already done from %s", c.Name)
 		return
-	} else if c.UsePriorTest != nil && *c.UsePriorTest {
-		prior := c.GetPrior("")
-		if prior != nil && !prior.Done() {
-			c.GetTest().Logf("trying to run prior %s", prior.Name)
-			parent := c.GetParent()
-			if parent == nil {
-				c.Fatalf("unable to run prior test %s because no parent", prior.Name)
-			}
-			c.GetTest().Run(prior.Name, func(u *testing.T) {
-				prior.SetTest(u, c.GetTest())
-				b.ExecuteOne(prior)
-			})
-		}
 	}
 
+	b.checkPriorTest(c)
+
 	// Do URL replacements
-	url, err := StringReplace(c, c.URL)
-	if err != nil {
-		c.Errorf("StringReplace failed: %v", err)
-	}
-	updatedURL, err := b.updateQueryString(c, url)
-	if err != nil {
-		c.Errorf("error updating query string: %v", err)
-	}
-	c.URL = updatedURL
+	c.urlReplace()
 
 	if !strings.HasPrefix(c.URL, "http:") && !strings.HasPrefix(c.URL, "https:") {
 		c.URL = c.GetDefaultURLBase() + c.URL
@@ -156,33 +82,8 @@ func (b *BaseClient) Do(c *Case) {
 	}
 
 	// Update request headers
-	updatedHeaders := map[string]string{}
-	for k, v := range c.RequestHeaders {
-		newK, err := StringReplace(c, k)
-		if err != nil {
-			c.Errorf("StringReplace for header %s failed: %v", k, err)
-			updatedHeaders[k] = v
-			continue
-		}
-		newV, err := StringReplace(c, v)
-		if err != nil {
-			c.Errorf("StringReplace for header value %s failed: %v", v, err)
-			updatedHeaders[newK] = v
-			continue
-		}
-		rq.Header.Set(newK, newV)
-		updatedHeaders[newK] = newV
-	}
-	c.RequestHeaders = updatedHeaders
-
-	if c.Verbose {
-		// TODO: Test for textual content-type header to set body true or false.
-		dump, err := httputil.DumpRequestOut(rq, true)
-		if err != nil {
-			c.GetTest().Logf("unable to dump request: %v", err)
-		}
-		fmt.Printf("%s\n", strings.ReplaceAll(string(dump), "\n", "\n> "))
-	}
+	c.RequestHeaders = c.updateRequestHeaders(rq)
+	c.dumpRequest(rq)
 
 	resp, err := b.Client.Do(rq)
 	if err != nil {
@@ -194,19 +95,9 @@ func (b *BaseClient) Do(c *Case) {
 		}
 	}()
 
-	if c.Verbose {
-		// TODO: Test for textual content-type header to set body true or false.
-		dump, err := httputil.DumpResponse(resp, true)
-		if err != nil {
-			c.GetTest().Logf("unable to dump response: %v", err)
-		}
-		fmt.Printf("\n\n< %s", strings.ReplaceAll(string(dump), "\n", "\n< "))
-	}
+	c.dumpResponse(resp)
 
-	status := resp.StatusCode
-	if status != c.Status {
-		c.Errorf("Expecting status %d, got %d", c.Status, status)
-	}
+	c.assertStatus(resp)
 
 	// TODO: This could consume a lot of memory, but for now this is what
 	// we want for being able to refer back to prior tests.
@@ -221,20 +112,28 @@ func (b *BaseClient) Do(c *Case) {
 
 	// TODO: This returns, which we don't want, we want to continue, which means
 	// we need to pass the testing harness around more.
-	for _, handler := range responseHandlers {
-		// Wind body to start in case it is not there.
-		_, err = c.GetResponseBody().Seek(0, io.SeekStart)
-		if err != nil {
-			c.Fatalf("Unable to seek response body to start: %v", err)
-		}
-
-		handler := handler
-		handler.Assert(c)
-	}
+	c.assertHandlers()
 
 	if c.Xfail && !c.GetXFailure() {
 		c.SetDone()
 		c.GetTest().Fatalf("Test passed when expecting failure.")
+	}
+}
+
+func (b *BaseClient) checkPriorTest(c *Case) {
+	if c.UsePriorTest != nil && *c.UsePriorTest {
+		prior := c.GetPrior("")
+		if prior != nil && !prior.Done() {
+			c.GetTest().Logf("trying to run prior %s", prior.Name)
+			parent := c.GetParent()
+			if parent == nil {
+				c.Fatalf("unable to run prior test %s because no parent", prior.Name)
+			}
+			c.GetTest().Run(prior.Name, func(u *testing.T) {
+				prior.SetTest(u, c.GetTest())
+				b.ExecuteOne(prior)
+			})
+		}
 	}
 }
